+++ caseless_dictionary/caseless_attribute_dict.py
